arraysのデータをlocaldata配下に移動した
