+++ app/src/main/java/com/hammer/app/subwaysimulator/ui/CreateRecipeActivity.kt
@@ -6,14 +6,14 @@ import android.view.View
 import android.widget.RadioButton
 import android.widget.RadioGroup
 import android.widget.Toast
-import com.hammer.app.subwaysimulator.model.Breads
-import com.hammer.app.subwaysimulator.model.Dressings
+import com.hammer.app.subwaysimulator.localdata.Breads
+import com.hammer.app.subwaysimulator.localdata.Dressings
 import com.hammer.app.subwaysimulator.R
 import com.hammer.app.subwaysimulator.model.Recipe
-import com.hammer.app.subwaysimulator.model.amounts
-import com.hammer.app.subwaysimulator.model.breads
-import com.hammer.app.subwaysimulator.model.dressings
-import com.hammer.app.subwaysimulator.model.sandwiches
+import com.hammer.app.subwaysimulator.localdata.amounts
+import com.hammer.app.subwaysimulator.localdata.breads
+import com.hammer.app.subwaysimulator.localdata.dressings
+import com.hammer.app.subwaysimulator.localdata.sandwiches
 import kotlinx.android.synthetic.main.create_recipe.*
 import kotlinx.android.synthetic.main.select_dressing_item.*
 import java.text.SimpleDateFormat
