+++ app/src/main/java/com/hammer/app/subwaysimulator/ui/EditRecipeActivity.kt
@@ -8,19 +8,19 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.*
-import com.hammer.app.subwaysimulator.model.Amounts
-import com.hammer.app.subwaysimulator.model.AmountsDressing
-import com.hammer.app.subwaysimulator.model.Breads
-import com.hammer.app.subwaysimulator.model.Dressings
+import com.hammer.app.subwaysimulator.localdata.Amounts
+import com.hammer.app.subwaysimulator.localdata.AmountsDressing
+import com.hammer.app.subwaysimulator.localdata.Breads
+import com.hammer.app.subwaysimulator.localdata.Dressings
 import com.hammer.app.subwaysimulator.R
 import com.hammer.app.subwaysimulator.model.Recipe
-import com.hammer.app.subwaysimulator.model.Sandwiches
-import com.hammer.app.subwaysimulator.model.amounts
-import com.hammer.app.subwaysimulator.model.amountsDressing
-import com.hammer.app.subwaysimulator.model.breads
-import com.hammer.app.subwaysimulator.model.dressings
-import com.hammer.app.subwaysimulator.model.dressingsWoNothing
-import com.hammer.app.subwaysimulator.model.sandwiches
+import com.hammer.app.subwaysimulator.localdata.Sandwiches
+import com.hammer.app.subwaysimulator.localdata.amounts
+import com.hammer.app.subwaysimulator.localdata.amountsDressing
+import com.hammer.app.subwaysimulator.localdata.breads
+import com.hammer.app.subwaysimulator.localdata.dressings
+import com.hammer.app.subwaysimulator.localdata.dressingsWoNothing
+import com.hammer.app.subwaysimulator.localdata.sandwiches
 import kotlinx.android.synthetic.main.create_recipe.*
 import kotlinx.android.synthetic.main.select_dressing_item.*
 import java.text.SimpleDateFormat
